requirements.txt
openai>=1.95.1
psycopg2-binary>=2.9.0
pgvector>=0.2.0
numpy>=1.24.0
python-dotenv>=1.0.0
mcp[cli]>=1.11.0
typing-extensions>=4.5.0
fastapi
pydantic
uvicorn
fastmcp